+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Changed
+- Updated to Enketo Core with new XPath evaluator.
 
 [1.12.2] - 2021-03-04
 ---------------------